finish from books

+++ server/database/queries/book/delete.ts
@@ -1,7 +1,7 @@
 import db from '../../config';
 
-const deleteBook = (bookName: string) => {
-  return db.promise().query('DELETE FROM book WHERE name = ?', [bookName]);
+const deleteBook = (bookName: string, userId: number) => {
+  return db.promise().query('DELETE FROM book WHERE name = ? AND userId = ?', [bookName, userId]);
 }
 
 export default deleteBook;
